Big one! I created user and auth models and subroutes (parent = /api).
Dependencies should be in packages.json but I also use mongoDB which
is installed globally so don't forget to npm install that also.
I choosed this way because with it we have layered architecture.

Questions, comments, please do!

Going out for a beer. Will be back in a couple of hours
and I'll continue to work on this.